dts: 'AI Camera' uses releasing gated clock
The clock using by IMX500 in 'AI Camera' is gpio-gated. The GPIO used is
provided by an I2C-controlled rpi-rp2040-gpio-bridge on 'AI Camera'.
Both the IMX500 and gpio-bridge share a common regulator for their power
supply. Using 'gpio-gate-clock', the 'AI Camera' will never be
powered-down as the GPIO provided by the gpio-bridge is always claimed
by the clock driver. Switching to 'gpio-gate-clock-releasing' causes the
GPIO to be released by the clock driver when the clock is not needed.
This allows 'AI Camera' to be powered-down.

Signed-off-by: Richard Oliver <richard.oliver@raspberrypi.com>